Change datatype cost and expanses in database.
Changed datatype for cost and expanses to string. Sothat the input can be checked and later be inverted to float at the right place. Added a function to check the user input and another one to add zero and euro symbol to input. Also fixed some errors. For example the case, if someone tries to select via 0 the -1 item in a list.